CP-51585: Add baggage to SpanContext
Carry around baggage with every span context. In order to propagate it
across tracing boundaries, each carrier must be adapted to understand
and propagate each value it wishes to propagate.

In future, all tracing boundaries must be adapted to propagate entire
SpanContexts, otherwise there is a juggling of related pieces of
information in different places - which is error-prone.

Signed-off-by: Colin James <colin.barr@cloud.com>