CP-51585: Forward baggage from xe-cli
Forwards baggage in the format expected by xapi-cli-server. Currently,
providing an environment variable (keyed "BAGGAGE") is the only way to
provide baggage to xe. In future, we may wish to extend this to allow
specifying baggage on the command-line, as arguments. This is not done
here due to unfamiliarity with how to parse key-value pairs on the
command-line with the current setup.

Signed-off-by: Colin James <colin.barr@cloud.com>