CP-52997: Track tgroups using `Tgroup.ThreadGroups`
Adds new module `Tgroup.ThreadGroups` that helps managing threads in
bucket. The purpose of this is to facilitate the calculation for the
scheduling of threads in batches instead of per thread.

The intention is iterating over 2-5 classes of threads every timeslice
for calculating the ideal time of individual threads by keeping track
of the total number of threads with each respective classification.

Signed-off-by: Gabriel Buica <danutgabriel.buica@cloud.com>